+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - New #25: Add `forwardMessage` and `forwardMessages` methods, and `MessageId` type.
 - New #26: Add `copyMessage` and `copyMessages` methods.
 - New #27: Add `sendAnimation` method.
+- New #28: Add `sendVoice` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.